fix(teststack): fix lint issues, zizmor injection, and missing env vars
- exitAfterDefer: call cancel() explicitly before os.Exit instead of
  deferring it (defers don't run on os.Exit)
- unconvert: sa.Id is already int64, remove redundant conversion
- unused: remove bearerTransport/bearerClient which were placeholder
  code that never got wired up; k6 uses http.DefaultClient directly
- zizmor template-injection: pass workflow_dispatch inputs through env
  vars in cloud-test-stack-cleanup.yml instead of inline template
  expansion in run: blocks
- missing env vars: CheckCloudInstanceTestsEnabled checks for
  GRAFANA_CLOUD_PROVIDER_AWS_ROLE_ARN, GRAFANA_CLOUD_PROVIDER_TEST_STACK_ID,
  and GRAFANA_CLOUD_PROVIDER_TEST_AWS_ACCOUNT_RESOURCE_ID unconditionally;
  add placeholder values so non-cloudprovider shards (grafana, slo, etc.)
  don't fail the env var assertion before they can skip

+++ tools/teststack/stack.go
@@ -123,7 +123,7 @@ func createAdminSA(ctx context.Context, client *gcom.APIClient, slug, name strin
 		Name:          name + "-token",
 		SecondsToLive: int32Ptr(int32((24 * time.Hour).Seconds())),
 	}
-	tok, _, err := client.InstancesAPI.PostInstanceServiceAccountTokens(ctx, slug, strconv.FormatInt(int64(*sa.Id), 10)).
+	tok, _, err := client.InstancesAPI.PostInstanceServiceAccountTokens(ctx, slug, strconv.FormatInt(*sa.Id, 10)).
 		PostInstanceServiceAccountTokensRequest(tokReq).
 		XRequestId(requestID()).
 		Execute()
@@ -131,7 +131,7 @@ func createAdminSA(ctx context.Context, client *gcom.APIClient, slug, name strin
 		return 0, "", fmt.Errorf("create admin SA token on %q: %w", slug, gcomErr(err))
 	}
 
-	return int64(*sa.Id), *tok.Key, nil
+	return *sa.Id, *tok.Key, nil
 }
 
 // deleteStack soft-deletes the stack. Idempotent: returns nil if the stack is
